+++ src/schemaRendering/schemaElements/Checkbox.js
