+++ src/schemaRendering/utils/conditionEvaluator.js
